api: v1alpha1: absorb normalize
with the addition of newer APIs, the normalization functions
are better served inside the API package. So let's move them
with their tests, fix the calling site(s) and remove the
normalize package.
